+++ core-ng/src/main/java/core/framework/api/kafka/Message.java
@@ -1,0 +1,9 @@
+package core.framework.api.kafka;
+
+/**
+ * @author neo
+ */
+public class Message<T> {
+    public String key;
+    public T value;
+}