+++ core-ng/src/main/java/core/framework/api/kafka/MessagePublisher.java
@@ -1,0 +1,10 @@
+package core.framework.api.kafka;
+
+/**
+ * @author neo
+ */
+public interface MessagePublisher<T> {
+    void publish(String key, T value);
+
+    void publish(String topic, String key, T value);
+}