+++ core-ng/src/main/java/core/framework/api/mongo/MongoCollection.java
@@ -20,6 +20,8 @@ public interface MongoCollection<T> {
 
     void insert(T entity);
 
+    void bulkInsert(List<T> entities);
+
     Optional<T> get(Get get);
 
     default Optional<T> get(Object id) {