+++ core-ng/src/main/java/core/framework/impl/kafka/MessageClassValidator.java
@@ -1,0 +1,12 @@
+package core.framework.impl.kafka;
+
+import core.framework.impl.validate.type.JAXBTypeValidator;
+
+/**
+ * @author neo
+ */
+final class MessageClassValidator extends JAXBTypeValidator {
+    MessageClassValidator(Class<?> messageClass) {
+        super(messageClass);
+    }
+}