+++ demo-service/src/main/java/app/queue/TestMessage.java
@@ -1,0 +1,17 @@
+package app.queue;
+
+import javax.xml.bind.annotation.XmlAccessType;
+import javax.xml.bind.annotation.XmlAccessorType;
+import javax.xml.bind.annotation.XmlElement;
+
+/**
+ * @author neo
+ */
+@XmlAccessorType(XmlAccessType.FIELD)
+public class TestMessage {
+    @XmlElement(name = "name")
+    public String name;
+
+    @XmlElement(name = "age")
+    public Integer age;
+}